--- src/app/course/[courseId]/layout.tsx
@@ -1,20 +1,0 @@
-import OfficeHourProvider from "@context/OfficeHourContext";
-
-interface LayoutProps {
-  children?: React.ReactNode;
-  params: {
-    courseId: string;
-  };
-}
-
-const Layout = (props: LayoutProps) => {
-  const {
-    children,
-    params: { courseId },
-  } = props;
-  return (
-    <OfficeHourProvider courseId={courseId}>{children}</OfficeHourProvider>
-  );
-};
-
-export default Layout;
