--- src/app/course/[courseId]/page.tsx
@@ -1,17 +1,0 @@
-import OfficeHourProvider from "@context/OfficeHourContext";
-
-interface PageProps {
-  params: {
-    courseId: string;
-  };
-}
-
-const Page = (props: PageProps) => {
-  const {
-    params: { courseId },
-  } = props;
-
-  return null;
-};
-
-export default Page;
