+++ src/app/private/course/[courseId]/page.tsx
@@ -1,5 +1,5 @@
 import Header from "@components/Header";
-import OfficeHourProvider from "../../../context/OfficeHourContext";
+import OfficeHourProvider from "@context/OfficeHourContext";
 import { IconButton } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
 interface PageProps {
@@ -19,7 +19,7 @@ const Page = (props: PageProps) => {
       <div>
         <Header
           leftIcon={
-            <IconButton>
+            <IconButton edge="start">
               <MenuIcon />
             </IconButton>
           }
